getBoo
