+++ LiLF/lib_util.py
@@ -3,7 +3,6 @@ import socket
 import datetime
 import tempfile
 
-from casacore import tables
 import numpy as np
 import multiprocessing, subprocess
 from threading import Thread
@@ -17,151 +16,12 @@ import warnings
 from astropy.io.fits.verify import VerifyWarning
 warnings.simplefilter('ignore', category=VerifyWarning)
 
-if (sys.version_info > (3, 0)):
-    from configparser import ConfigParser
-else:
-    from ConfigParser import ConfigParser
-
 # load here to be sure to have "Agg" at the beginning
 import matplotlib as mpl
 mpl.use("Agg")
 
 from LiLF import lib_img
 from LiLF.lib_log import logger
-
-def getParset(parsetFile=''):
-    """
-    Get parset file and return dict of values
-    """
-    def add_default(section, option, val):
-        if not config.has_option(section, option): config.set(section, option, val)
-    
-    if parsetFile == '':
-        matched_conf_files = glob.glob('[Ll][Ii][Ll][Ff].conf*') + glob.glob('../[Ll][Ii][Ll][Ff].conf*')
-        if len(matched_conf_files) > 1:
-            raise LookupError(f'Found more than one configuration file: {matched_conf_files}')
-        elif len(matched_conf_files) == 1:
-            parsetFile = matched_conf_files[0]
-            logger.info(f'Found config file: {parsetFile}')
-
-    config = ConfigParser(defaults=None)
-    config.read(parsetFile)
-    
-    # add pipeline sections and defaul parset dir:
-    for pipeline in glob.glob(os.path.dirname(__file__)+'/../parsets/*'):
-        pipeline = os.path.basename(pipeline)
-        if not config.has_section(pipeline): config.add_section(pipeline)
-        if not config.has_option(pipeline, 'parset_dir'):
-                config.set(pipeline, 'parset_dir', os.path.dirname(__file__)+'/../parsets/'+pipeline)
-    # add other sections
-    if not config.has_section('flag'): config.add_section('flag')
-    if not config.has_section('model'): config.add_section('model')
-    if not config.has_section('PiLL'): config.add_section('PiLL')
-
-    ### LOFAR ###
-
-    # PiLL
-    add_default('PiLL', 'working_dir', os.getcwd())
-    add_default('PiLL', 'redo_cal', 'False') # re-do the calibrator although it is in the archive
-    add_default('PiLL', 'download_file', '') # html.txt file to use instead of staging
-    add_default('PiLL', 'project', '')
-    add_default('PiLL', 'target', '')
-    add_default('PiLL', 'obsid', '') # unique ID
-    add_default('PiLL', 'minmaxhrs', '0,9999') # min and max hours for an obs to be selected
-    add_default('PiLL', 'logfile', '') # logfile for PiLL
-    # preprocess
-    add_default('LOFAR_preprocess', 'fix_table', 'True') # fix bug in some old observations
-    add_default('LOFAR_preprocess', 'renameavg', 'True')
-    add_default('LOFAR_preprocess', 'keep_IS', 'True')
-    add_default('LOFAR_preprocess', 'backup_full_res', 'False')
-    add_default('LOFAR_preprocess', 'demix_sources', '')  # Demix  sources in these patches (e.g. [VirA,TauA], default: No demix
-    add_default('LOFAR_preprocess', 'demix_skymodel', '')  # Use non-default demix skymodel.
-    add_default('LOFAR_preprocess', 'demix_field_skymodel', 'LOTSS-DR3')  # Provide a custom target skymodel instead of online gsm model. Set to '' to ignore target.
-    add_default('LOFAR_preprocess', 'run_aoflagger', 'False')  # run aoflagger on individual sub-bands, only in cases where this was not one by the observatory!
-    add_default('LOFAR_preprocess', 'tar', 'True')  # Tar MS files at the end
-    add_default('LOFAR_preprocess', 'data_dir', '')
-    # cal
-    add_default('LOFAR_cal', 'data_dir', 'data-bkp/')
-    add_default('LOFAR_cal', 'skymodel', '') # by default use calib-simple.skydb for LBA and calib-hba.skydb for HBA
-    add_default('LOFAR_cal', 'imaging', 'False')
-    add_default('LOFAR_cal', 'fillmissingedges', 'True')
-    add_default('LOFAR_cal', 'less_aggressive_flag', 'False') # change flagging so that we can handle data with alternating SBs only or many flagged points
-    add_default('LOFAR_cal', 'develop', 'False') # if true prevents the deletion of files
-    add_default('LOFAR_cal', 'use_GNSS', 'False') # Use GNSS satellite data for pre correcting TEC and FR
-    add_default('LOFAR_cal', 'use_shm', 'False') # use /dev/shm for temporary files, if available
-    # add_default('LOFAR_cal', 'beam_model', 'hamaker') #
-    # timesplit
-    add_default('LOFAR_timesplit', 'data_dir', 'data-bkp/')
-    add_default('LOFAR_timesplit', 'cal_dir', '') # by default the repository is tested, otherwise ../obsid_3[c|C]*
-    add_default('LOFAR_timesplit', 'ngroups', '1')
-    add_default('LOFAR_timesplit', 'initc', '0')
-    add_default('LOFAR_timesplit', 'fillmissingedges', 'True')
-    add_default('LOFAR_timesplit', 'apply_fr', 'False') # Also transfer rotationmeasure sols? (E.g. for nearby calibrator and target)
-    add_default('LOFAR_timesplit', 'no_aoflagger', 'False') # TEST: Skip aoflagger (e.g. for observations of A-Team sources)
-    add_default('LOFAR_timesplit', 'ateam_clip', '') # [CygA, CasA] or [CasA] or [CygA] or '' - the code clips the specified ateams (if not demixed from the observatory)
-    add_default('LOFAR_timesplit', 'use_GNSS', 'False') # Use GNSS satellite data for pre correcting TEC and FR
-    # ddparallel
-    add_default('LOFAR_ddparallel', 'maxIter', '2')
-    add_default('LOFAR_ddparallel', 'subfield', '') # possible to provide a ds9 box region customized sub-field. DEfault='' -> Automated detection using subfield_min_flux.
-    add_default('LOFAR_ddparallel', 'subfield_min_flux', '20') # min flux within calibration subfield
-    add_default('LOFAR_ddparallel', 'ph_sol_mode', 'phase') # phase or tecandphase
-    add_default('LOFAR_ddparallel', 'remove3c', 'True')
-    add_default('LOFAR_ddparallel', 'fulljones', 'False')
-    add_default('LOFAR_ddparallel', 'min_facets', '')
-    add_default('LOFAR_ddparallel', 'max_facets', '')
-    add_default('LOFAR_ddparallel', 'develop', 'False') # if true make more debug images (slower)
-    add_default('LOFAR_ddparallel', 'data_dir', '')
-    add_default('LOFAR_ddparallel', 'use_shm', 'False') # use /dev/shm for temporary files, if available
-    # ddserial
-    add_default('LOFAR_ddserial', 'maxIter', '1')
-    add_default('LOFAR_ddserial', 'minCalFlux60', '0.8')
-    add_default('LOFAR_ddserial', 'solve_amp', 'True') # to disable amp sols
-    add_default('LOFAR_ddserial', 'use_shm', 'False') # use /dev/shm for temporary files, if available
-    add_default('LOFAR_ddserial', 'use_shm_ddcal', 'True') # use /dev/shm for temporary ddcal files, if available
-    add_default('LOFAR_ddserial', 'target_dir', '') # ra,dec
-    add_default('LOFAR_ddserial', 'manual_dd_cal', '')
-    add_default('LOFAR_ddserial', 'develop', 'False') # if true make more debug images (slower)
-    # add_default('LOFAR_ddserial', 'solve_tec', 'False') # per default, solve each dd for scalarphase. if solve_tec==True, solve for TEC instead.
-    # extract
-    add_default('LOFAR_extract', 'max_niter', '10')
-    add_default('LOFAR_extract', 'subtract_region', '') # Sources inside extract-reg that should still be subtracted! Use this e.g. for individual problematic sources in a large extractReg
-    add_default('LOFAR_extract', 'ph_sol_mode', 'phase') # tecandphase, phase
-    add_default('LOFAR_extract', 'amp_sol_mode', 'diagonal') # diagonal, fulljones
-    add_default('LOFAR_extract', 'beam_cut', '0.3') # up to which distance a pointing will be considered
-    add_default('LOFAR_extract', 'no_selfcal', 'False') # just extract the data, do not perform selfcal - use this if u want to use e.g. Reinout van Weeren's facet_seflcal script
-    add_default('LOFAR_extract', 'ampcal', 'auto')
-    add_default('LOFAR_extract', 'extractRegion', 'target.reg')
-    # quality
-    add_default('LOFAR_quality', 'ddparallel_dir', 'ddparallel')
-    add_default('LOFAR_quality', 'ddserial_dir', 'ddserial')
-    # virgo
-    add_default('LOFAR_virgo', 'cal_dir', '')
-    add_default('LOFAR_virgo', 'data_dir', './')
-    # m87
-    add_default('LOFAR_m87', 'data_dir', './')
-    add_default('LOFAR_m87', 'updateweights', 'False')
-    add_default('LOFAR_m87', 'skipmodel', 'False')
-    add_default('LOFAR_m87', 'model_dir', '')
-    # peel
-    ### uGMRT ###
-    # init - deprecated
-    #add_default('uGMRT_init', 'data_dir', './datadir')
-    # cal - deprecated
-    #add_default('uGMRT_cal', 'skymodel', os.path.dirname(__file__)+'/../models/calib-simple.skydb')
-
-    ### General ###
-
-    # flag
-    add_default('flag', 'stations', '') # LOFAR
-    add_default('flag', 'antennas', '') # uGMRT
-    # model
-    add_default('model', 'sourcedb', '')
-    add_default('model', 'fits_model', '')
-    add_default('model', 'apparent', 'False')
-    add_default('model', 'userReg', '')
-
-
-    return config
 
 def create_extregion(ra, dec, extent, color='yellow'):
     """
@@ -529,324 +389,6 @@ class Region_helper():
 
     def __len__(self):
         return len(self.reg_list)
-
-class Skip(Exception):
-    pass
-
-class Exit(Exception):
-    pass
-
-class Walker():
-    """
-    An object of this class may be used to re-run a pipeline without repeating steps that were completed previously.
-    Use like:
-    w = Walker("filename.walker")
-    with w.if_todo("stepname"):
-        Do whatever...
-
-    Adopted from https://stackoverflow.com/questions/12594148/skipping-execution-of-with-block
-    """
-    def __init__(self, filename):
-        open(filename, 'a').close() # create the file if doesn't exists
-        self.filename = os.path.abspath(filename)
-        self.__skip__ = False
-        self.__step__ = None
-        self.__inittime__ = None
-        self.__globaltimeinit__ = datetime.datetime.now()
-
-    def if_todo(self, stepname):
-        """
-        This is basically a way to get a context manager to accept an argument. Will return "self" as context manager
-        if called as context manager.
-        """
-        self.__skip__ = False
-        self.__step__ = stepname
-        with open(self.filename, "r") as f:
-            for stepname_done in f:
-                if stepname == stepname_done.split('#')[0].rstrip():
-                    self.__skip__ = True
-        return self
-
-    def __enter__(self):
-        """
-        Skips body of with-statement if __skip__.
-        This uses some kind of dirty hack that might only work in CPython.
-        """
-        if self.__skip__:
-            sys.settrace(lambda *args, **keys: None)
-            frame = sys._getframe(1)
-            frame.f_trace = self.trace
-        else:
-            logger.log(20, '>> start >> {}'.format(self.__step__))
-            self.__timeinit__ = datetime.datetime.now()
-
-    def trace(self, frame, event, arg):
-        raise Skip()
-
-    def __exit__(self, type, value, traceback):
-        """
-        Catch "Skip" errors, if not skipped, write to file after exited without exceptions.
-        """
-        if type is None:
-            with open(self.filename, "a") as f:
-                delta_td = datetime.datetime.now() - self.__timeinit__
-                total_seconds = int(delta_td.total_seconds())
-                days, rem = divmod(total_seconds, 86400)
-                hours, rem = divmod(rem, 3600)
-                minutes, seconds = divmod(rem, 60)
-                delta = f"{days}d {hours}h {minutes}m {seconds}s"
-                f.write(f"{self.__step__} # {delta}\n")
-            logger.info('<< done << {}'.format(self.__step__))
-            return  # No exception
-        if issubclass(type, Skip):
-            logger.warning('>> skip << {}'.format(self.__step__))
-            return True  # Suppress special SkipWithBlock exception
-        if issubclass(type, Exit):
-            logger.error('<< exit << {}'.format(self.__step__))
-            return True
-
-    def alldone(self):
-        delta = 'h '.join(str(datetime.datetime.now() - self.__globaltimeinit__).split(':')[:-1])+'m'
-        logger.info('Done. Total time: '+delta)
-
-class Scheduler():
-    def __init__(self, maxProcs = None, max_cpucores = None, log_dir = 'logs', dry = False):
-        """
-        maxProcs:       max number of parallel processes
-        max_cpucores:   max number of cpu cores usable in a node
-        dry:            don't schedule job
-        """
-        self.hostname = socket.gethostname()
-        self.cluster = self.get_cluster()
-        self.log_dir = log_dir
-        #self.qsub    = qsub
-        # if qsub/max_thread/max_cpucores not set, guess from the cluster
-        # if they are set, double check number are reasonable
-        #if (self.qsub == None):
-        #    self.qsub = False
-        #else:
-        #    if ((self.qsub is False and self.cluster == "Hamburg") or
-        #       (self.qsub is True and (self.cluster == "Leiden" or self.cluster == "CEP3" or
-        #                               self.cluster == "Hamburg_fat" or self.cluster == "Pleiadi" or self.cluster == "Herts"))):
-        #        logger.critical('Qsub set to %s and cluster is %s.' % (str(qsub), self.cluster))
-        #        sys.exit(1)
-
-        if (max_cpucores == None):
-            # check if running in a slurm environment with a limited number of CPUs (less than cpu_count())
-            slurm_cpus = os.getenv('SLURM_CPUS_ON_NODE', False)
-            if slurm_cpus:
-                self.max_cpucores = int(slurm_cpus)
-            else:
-                self.max_cpucores = multiprocessing.cpu_count()
-        else:
-            self.max_cpucores = max_cpucores
-
-        if (maxProcs is None) or (maxProcs > self.max_cpucores):
-            self.maxProcs = self.max_cpucores
-        else:
-            self.maxProcs = maxProcs
-
-        self.dry = dry
-
-        logger.info("Scheduler initialised for cluster " + self.cluster + ": " + self.hostname +
-                    " (maxProcs: " + str(self.maxProcs) + ", max_cpucores: " + str(self.max_cpucores) + ").")
-
-        self.action_list = []
-        self.log_list    = []  # list of 2-tuples of the type: (log filename, type of action)
-
-
-    def get_cluster(self):
-        """
-        Find in which computing cluster the pipeline is running
-        """
-        hostname = self.hostname
-        if (hostname == 'lgc1' or hostname == 'lgc2'):
-            return "Hamburg"
-        elif ('r' == hostname[0] and 'c' == hostname[3] and 's' == hostname[6]):
-            return "Pleiadi"
-        elif ('node3' in hostname):
-            return "Hamburg_fat"
-        elif ('node' in hostname):
-            return "Herts"
-        elif ('leidenuniv' in hostname):
-            return "Leiden"
-        elif ('spider' in hostname):
-            return "Spider"
-        else:
-            logger.debug('Hostname %s unknown.' % hostname)
-            return "Unknown"
-
-
-    def add(self, cmd = '', log = '', logAppend = True, commandType = ''):
-        """
-        Add a command to the scheduler list
-        cmd:         the command to run
-        log:         log file name that can be checked at the end
-        logAppend:   if True append, otherwise replace
-        commandType: can be a list of known command types as "wsclean", "DP3", ...
-        """
-
-        if (log != ''):
-            log = self.log_dir + '/' + log
-
-            if (logAppend):
-                cmd += " >> "
-            else:
-                cmd += " > "
-            cmd += log + " 2>&1"
-
-        if commandType == 'wsclean':
-            logger.debug('Running wsclean: %s' % cmd)
-        elif commandType == 'DP3':
-            logger.debug('Running DP3: %s' % cmd)
-        #elif commandType == 'singularity':
-        #    cmd = 'SINGULARITY_TMPDIR=/dev/shm singularity exec -B /tmp,/dev/shm,/localwork,/localwork.ssd,/home /home/fdg/node31/opt/src/lofar_sksp_ddf.simg ' + cmd
-        #    logger.debug('Running singularity: %s' % cmd)
-        elif (commandType.lower() == "ddfacet" or commandType.lower() == 'ddf'):
-            logger.debug('Running DDFacet: %s' % cmd)
-        elif commandType == 'python':
-            logger.debug('Running python: %s' % cmd)
-        else:
-            logger.debug('Running general: %s' % cmd)
-
-        #if self.qsub:
-        #    if qsub_cpucores == 'max':
-        #        qsub_cpucores = self.max_cpucores
-        #    # if number of cores not specified, try to find automatically
-        #    elif qsub_cpucores == None:
-        #        qsub_cpucores = 1 # default use single CPU
-        #        if ("DP3" == cmd[ : 4]):
-        #            qsub_cpucores = 1
-        #        if ("wsclean" == cmd[ : 7]):
-        #            qsub_cpucores = self.max_cpucores
-        #    if (qsub_cpucores > self.max_cpucores):
-        #        qsub_cpucores = self.max_cpucores
-        #    self.action_list.append([str(qsub_cpucores), '\'' + cmd + '\''])
-        #else:
-        self.action_list.append(cmd)
-
-        if (log != ""):
-            self.log_list.append((log, commandType))
-
-
-    def run(self, check = False, maxProcs = None):
-        """
-        If 'check' is True, a check is done on every log in 'self.log_list'.
-        If max_thread != None, then it overrides the global values, useful for special commands that need a lower number of threads.
-        """
-
-        def worker(queue):
-            for cmd in iter(queue.get, None):
-                #if self.qsub and self.cluster == "Hamburg":
-                #    cmd = 'salloc --job-name LBApipe --time=24:00:00 --nodes=1 --tasks-per-node='+cmd[0]+\
-                #            ' /usr/bin/srun --ntasks=1 --nodes=1 --preserve-env \''+cmd[1]+'\''
-                gc.collect()
-                subprocess.call(cmd, shell = True)
-
-        # limit number of processes
-        if (maxProcs == None):
-            maxProcs_run = self.maxProcs
-        else:
-            maxProcs_run = min(maxProcs, self.maxProcs)
-
-        q       = Queue()
-        threads = [Thread(target = worker, args=(q,)) for _ in range(maxProcs_run)]
-
-        for i, t in enumerate(threads): # start workers
-            t.daemon = True
-            t.start()
-
-        for action in self.action_list:
-            if (self.dry):
-                continue # don't schedule if dry run
-            q.put_nowait(action)
-        for _ in threads:
-            q.put(None) # signal no more commands
-        for t in threads:
-            t.join()
-
-        # check outcomes on logs
-        if (check):
-            for log, commandType in self.log_list:
-                self.check_run(log, commandType)
-
-        # reset list of commands
-        self.action_list = []
-        self.log_list    = []
-
-
-    def check_run(self, log = "", commandType = ""):
-        """
-        Produce a warning if a command didn't close the log properly i.e. it crashed
-        NOTE: grep, -L inverse match, -l return only filename
-        """
-
-        if (not os.path.exists(log)):
-            logger.warning("No log file found to check results: " + log)
-            return 1
-
-        if (commandType == "DP3"):
-            out = subprocess.check_output(r'grep -L "Finishing processing" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Segmentation fault\|Killed" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Aborted (core dumped)" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -i -l "Exception" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            #out += subprocess.check_output(r'grep -i -l "already a beam correction applied" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            # this interferes with the missingantennabehaviour=error option...
-            # out += subprocess.check_output('grep -l "error" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "misspelled" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-        elif (commandType == "CASA"):
-            out = subprocess.check_output(r'grep -l "[a-z]Error" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "An error occurred running" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "\*\*\* Error \*\*\*" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-        elif (commandType == "wsclean"):
-            out = subprocess.check_output(r'grep -l "exception occur" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Segmentation fault\|Killed" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Aborted" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Bus error" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "(core dumped)" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            # out += subprocess.check_output('grep -L "Cleaning up temporary files..." '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-        elif (commandType.lower() == "ddfacet" or commandType.lower() == 'ddf'):
-            out = subprocess.check_output(r'grep -l "Traceback (most recent call last):" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "exception occur" ' + log + ' ; exit 0', shell=True, stderr=subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "raise Exception" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Segmentation fault\|Killed" ' + log + ' ; exit 0', shell=True,
-                                           stderr=subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "killed by signal" ' + log + ' ; exit 0', shell=True,
-                                           stderr=subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Aborted" ' + log + ' ; exit 0', shell=True, stderr=subprocess.STDOUT)
-
-        elif (commandType == "python"):
-            out = subprocess.check_output(r'grep -l "Traceback (most recent call last):" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Segmentation fault\|Killed" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            # out += subprocess.check_output(r'grep -i -l \'(?=^((?!error000).)*$).*Error.*\' '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -i -l "Critical" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "ERROR" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "ImportError" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Traceback (most recent call last)" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-            out += subprocess.check_output(r'grep -l "Permission denied" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-#        elif (commandType == "singularity"):
-#            out = subprocess.check_output('grep -l "Traceback (most recent call last):" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-#            out += subprocess.check_output('grep -i -l \'(?=^((?!error000).)*$).*Error.*\' '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-#            out += subprocess.check_output('grep -i -l "Critical" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-        elif (commandType == "general"):
-            out = subprocess.check_output('grep -l -i "error" '+log+' ; exit 0', shell = True, stderr = subprocess.STDOUT)
-
-        else:
-            logger.warning("Unknown command type for log checking: '" + commandType + "'")
-            return 1
-
-        if out != b'':
-            out = out.split(b'\n')[0].decode()
-            logger.error(commandType+' run problem on:\n'+out)
-            errlines = subprocess.check_output('tail -n 10 '+log, shell = True, stderr = subprocess.STDOUT).decode()
-            raise RuntimeError(commandType+' run problem on:\n'+out+'\n'+errlines)
-
-        return 0
 
 def get_template_image(reference_ra_deg, reference_dec_deg, ximsize=512, yimsize=512, cellsize_deg=0.000417, fill_val=0):
     """
