+++ pipelines/PiLL.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import os, sys, glob, getpass, socket, pickle
-from LiLF.surveys_db import SurveysDB
 from LiLF import lib_util, lib_log
 logger_obj = lib_log.Logger('PiLL')
 logger = lib_log.logger
@@ -21,17 +20,6 @@ obsid = parset.get('PiLL','obsid')
 download_file = parset.get('PiLL','download_file')
 if download_file != '': download_file = os.path.abspath(download_file)
 
-def calibrator_tables_available(obsid):
-    """
-    check if calibrator data exist in the database
-    """
-    with SurveysDB(survey='lba',readonly=True) as sdb:
-        sdb.execute('SELECT * FROM observations WHERE id=%f' % obsid)
-        r = sdb.cur.fetchall()
-        if len(r) != 0 and r[0]['location'] != '': return True
-        else: return False
-
-
 def local_calibrator_dirs(searchdir='', obsid=None):
     """
     Return the dirname of the calibrators
@@ -49,11 +37,6 @@ def local_calibrator_dirs(searchdir='', obsid=None):
     else: return calibrators
 
 
-def update_status_db(field, status):
-    with SurveysDB(survey='lba',readonly=False) as sdb:
-        r = sdb.execute('UPDATE fields SET status="%s" WHERE id="%s"' % (status,field))
-
-
 def check_done(pipename):
     """
     check if "Done" is written in the last line of the log file, otherwise quit with error.
@@ -62,7 +45,6 @@ def check_done(pipename):
     with open(logfile, 'r') as f:
         last_line = f.readlines()[-1]
     if not "Done" in last_line:
-        if survey: update_status_db(target, 'Error') 
         logger.error('Something went wrong in the last pipeline call.')
         sys.exit()
 
@@ -85,49 +67,6 @@ def fix_dirs(target):
         if c != '%s_-_%s' % (idx,cal):
             os.system('mv %s %s_-_%s' % (c,idx,cal))
 
-
-####################################################################################
-
-# If no input given, assume survey data and query the database for data to process
-survey = False
-if download_file == '' and project == '' and target == '' and obsid == '':
-    survey = True
-    if os.path.exists('target.txt'):
-        with open('target.txt', 'r') as file:
-                target,target_ra,target_dec = file.read().replace('\n', '').split(',')
-                target_ra = float(target_ra)
-                target_dec = float(target_dec)
-    else:
-        logger.info('### Quering database...')
-        with SurveysDB(survey='lba',readonly=True) as sdb:
-            sdb.execute('SELECT * FROM fields WHERE status="Observed" order by priority asc')
-            r = sdb.cur.fetchall()
-            if len(r) == 0:
-                logger.warning('No field left in the db...')
-                sys.exit()
-            target = r[0]['id'] # here we set $target
-            target_ra = r[0]['ra']
-            target_dec = r[0]['decl']
-        # save target name
-        with open("target.txt", "w") as file:
-                print('%s,%f,%f' % (target,target_ra,target_dec), file=file)
-
-    with SurveysDB(survey='lba',readonly=True) as sdb:
-        sdb.execute('SELECT * FROM field_obs WHERE field_id="%s"' % target)
-        r = sdb.cur.fetchall()
-        obsid = ','.join([str(x['obs_id']) for x in r]) # here we set $obsid
-
-    logger.info("### Working on target: %s (obsid: %s)" % (target, obsid))
-    # add other info, like cluster, node, user...
-    username = getpass.getuser()
-    clustername = s.cluster
-    nodename = socket.gethostname()
-    with SurveysDB(survey='lba',readonly=False) as sdb:
-        r = sdb.execute('UPDATE fields SET username="%s" WHERE id="%s"' % (username, target))
-        r = sdb.execute('UPDATE fields SET clustername="%s" WHERE id="%s"' % (clustername, target))
-        r = sdb.execute('UPDATE fields SET nodename="%s" WHERE id="%s"' % (nodename, target))
-
-if survey: update_status_db(target, 'Download') 
 
 #######
 # setup
@@ -153,12 +92,7 @@ with w.if_todo('download'):
     if download_file == '':
         cmd = LiLF_dir+'/scripts/LOFAR_stager.py --quiet --nocal'
         if target != '':
-            if survey:
-                #cmd += ' --target %s' % target[:-1] # remove the "o" or "s" at the end
-                # use coordinates
-                cmd += ' --radecdist %f,%f,0.1' % (target_ra, target_dec)
-            else:
-                cmd += ' --target %s' % target
+            cmd += ' --target %s' % target
         if obsid != '':
             cmd += ' --obsID %s' % obsid
         logger.debug("Exec: %s" % cmd)
@@ -175,8 +109,6 @@ if target != '': fix_dirs(target) # try to harmonize target dirs
 target_dirs = [t for t in glob.glob('id*') if t not in local_calibrator_dirs()]
 logger.debug('TARGET DIRS: %s' % (','.join(target_dirs) ) )
 
-if survey: update_status_db(target, 'Calibrator')
-
 for target_dir in target_dirs:
     ##########
     # calibrator
@@ -184,7 +116,7 @@ for target_dir in target_dirs:
     # then it also runs the calibrator pipeline
     obsid = int(target_dir.split('_-_')[0][2:])
     with w.if_todo('cal_id%i' % obsid):
-        if not survey or redo_cal or not calibrator_tables_available(obsid):
+        if redo_cal:
             logger.info('### %s: Starting calibrator... #####################################' % target_dir)
             cal_dir = local_calibrator_dirs(working_dir, obsid)
             # if calibrator not downloaded, do it
@@ -203,33 +135,6 @@ for target_dir in target_dirs:
                 os.system('mv *MS data-bkp')
             os.system(LiLF_dir+'/pipelines/LOFAR_cal.py')
             check_done('pipeline-cal')
-
-            if survey: # only backup solutions if survey
-                # copy solutions in the repository
-                cal_dir = os.path.basename(local_calibrator_dirs(working_dir, obsid)[0])
-                logger.info('Copy: cal*h5 -> herts:/beegfs/lofar/lba/calibration_solutions/%s' % cal_dir)
-                os.system('ssh herts "rm -rf /beegfs/lofar/lba/calibration_solutions/%s"' % cal_dir)
-                os.system('ssh herts "mkdir /beegfs/lofar/lba/calibration_solutions/%s"' % cal_dir)
-                os.system('scp -q cal-pa.h5 cal-bp.h5 cal-iono.h5 cal-iono-cs.h5 herts:/beegfs/lofar/lba/calibration_solutions/%s' % cal_dir)
-                os.system('scp -q -r plots* herts:/beegfs/lofar/lba/calibration_solutions/%s' % cal_dir)
-
-                # update the db
-                with SurveysDB(survey='lba',readonly=False) as sdb:
-                    sdb.execute('INSERT INTO observations (id,location,calibratordata) VALUES \
-                    (%i,"herts","%s")' % (obsid, "/beegfs/lofar/lba/calibration_solutions/"+cal_dir))
-
-        else:
-            # download calibrator solutions
-            with SurveysDB(survey='lba',readonly=True) as sdb:
-                sdb.execute('select location,calibratordata from observations where id=%i' % obsid)
-                r = sdb.cur.fetchall()[0]
-            location = r['location']
-            calibratordata = r['calibratordata']
-            os.chdir(working_dir)
-            cal_dir = calibratordata.split('/')[-1]
-            logger.info('Downloading solutions: %s:%s/cal-*h5' % (location,calibratordata))
-            os.system('mkdir %s' % cal_dir)
-            os.system('scp -q %s:%s/cal-*h5 %s' % (location,calibratordata,cal_dir))
     ### DONE
 
     ##########
@@ -266,7 +171,6 @@ for grouped_target in grouped_targets:
 
     # selfcal
     with w.if_todo('self_%s' % grouped_target):
-        if survey: update_status_db(grouped_target, 'Self')
         logger.info('### %s: Starting selfcal #####################################' % grouped_target)
         os.system(LiLF_dir+'/pipelines/LOFAR_self.py')
         check_done('pipeline-self')
@@ -274,46 +178,13 @@ for grouped_target in grouped_targets:
 
     # DD-cal
     with w.if_todo('dd_%s' % grouped_target):
-        if survey: update_status_db(grouped_target, 'Ddcal')
         logger.info('### %s: Starting ddcal #####################################' % grouped_target)
         os.system(LiLF_dir+'/pipelines/LOFAR_dd.py')
         check_done('pipeline-dd')
     ### DONE
 
-    if survey: # only back up solutions if survey
-        with w.if_todo('saveproducts_%s' % grouped_target):
-            if survey: update_status_db(grouped_target, 'SaveProducts')
-            # copy images in herts
-            logger.info('Copy ddcal products -> lofar.herts.ac.uk:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('ssh herts "rm -rf /beegfs/lofar/lba/products/%s"' % grouped_target)
-            os.system('ssh herts "mkdir /beegfs/lofar/lba/products/%s"' % grouped_target)
-            os.system('scp -q self/images/wideP*.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q self/images/wideM-1-MFS-image.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q self/images/wide-largescale-MFS-image.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('ssh herts "mkdir /beegfs/lofar/lba/products/%s/plots"' % grouped_target)
-            os.system('scp -q -r self/plots/* herts:/beegfs/lofar/lba/products/%s/plots' % grouped_target)
-            os.system('scp -q ddcal/c0*/images/wideDD-c*.MFS-image.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q ddcal/c0*/images/wideDD-c*.MFS-image-pb.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q ddcal/c0*/images/wideDD-c*.MFS-residual.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q ddcal/c01/solutions/interp.h5 herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q ddcal/c0*/skymodels/all*reg herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('scp -q ddcal/primarybeam.fits herts:/beegfs/lofar/lba/products/%s' % grouped_target)
-            os.system('ssh herts "mkdir /beegfs/lofar/lba/products/%s/init"' % grouped_target)
-            os.system('scp -q -r ddcal/init/*model.fits ddcal/init/wideM-1-sources.txt herts:/beegfs/lofar/lba/products/%s/init' % grouped_target)
-            # logs
-            os.system('ssh herts "mkdir /beegfs/lofar/lba/products/%s/logs"' % grouped_target)
-            os.system('scp -q ../*logger ../*walker ../*%s*/*logger ../*%s*/*walker herts:/beegfs/lofar/lba/products/%s/logs' % (grouped_target, grouped_target, grouped_target))
-            # copy ms in Bologna
-            logger.info('Copy mss -> pleiadi:/iranet/lofarfs2/lofar2/fdg/surveytgts/%s' % grouped_target)
-            os.system('tar zcf %s.tgz mss-avg' % grouped_target)
-            os.system('ssh pleiadi "rm -rf /iranet/lofarfs2/lofar2/fdg/surveytgts/%s"' % grouped_target)
-            os.system('ssh pleiadi "mkdir -p /iranet/lofarfs2/lofar2/fdg/surveytgts/%s"' % grouped_target)
-            os.system('scp -q %s.tgz pleiadi:/iranet/lofarfs2/lofar2/fdg/surveytgts/%s' % (grouped_target,grouped_target))
-        ### DONE
-
     # Quality check
     with w.if_todo('quality_%s' % grouped_target):
-        if survey: update_status_db(grouped_target, 'QualityCheck')
         logger.info('### %s: Starting quality check #####################################' % grouped_target)
         os.system(LiLF_dir+'/pipelines/LOFAR_quality.py')
         check_done('pipeline-quality')
@@ -326,11 +197,6 @@ for grouped_target in grouped_targets:
         logger.info('DDcal residual rms noise (cycle 1): %.1f mJy/b' % (qdict['ddcal_c1_rms'] * 1e3))
         logger.info('DDcal NVSS ratio (cycle 1): %.1f with %i matches' % (qdict['nvss_ratio'], qdict['nvss_match']))
         logger.info('DDcal total flags: %.1f%%' % (qdict['flag_frac']*100))
-        if survey:
-            with SurveysDB(survey='lba', readonly=False) as sdb:
-                r = sdb.execute('UPDATE fields SET noise="%s", nvss_ratio="%s", nvss_match="%s", flag_frac="%s" WHERE id="%s"' \
-                        % (qdict['ddcal_c1_rms'],qdict['nvss_ratio'], qdict['nvss_match'], qdict['flag_frac'],  grouped_target))
-    ### DONE
+     ### DONE
 
-    if survey: update_status_db(grouped_target, 'Done')
     logger.info('### %s: Done. #####################################' % grouped_target)
